trading_bot/requirements.txt
python-binance==1.0.19
python-dotenv==1.0.1

